Move add_explicit_area_ids to docparagraph.py
